LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing/ambiguous-ampersand-expected.txt

Link
Text: ?a=b&c;=d&a0b;=c©=1¬i;=n¬=in∉=∉¬&;& &

Link
Text: ?a=b&c=d&a0b=c©=1¬i=n¬=in¬in=∉¬&;& &


PASS Check number of divs
FAIL Check div structure: network assert_equals: text network expected "Text: ?a=b&c=d&a0b=c©=1¬i=n¬=in¬in=∉¬&;& &" but got "Text: ?a=b&c;=d&a0b;=c©=1¬i;=n¬=in∉=∉¬&;& &"
PASS Check div structure: document.write
